project1/requirements.txt
numpy
opencv-contrib-python
scikit-learn
pillow
matplotlib
tqdm
torchviz
nvitop
ipykernel
python-dotenv
ml_collections
scipy
gdown
ultralytics
timm
filterpy
Cython
optuna
pandas
einops
timm
wandb